+++ test/Test.js
@@ -1,18 +1,12 @@
-const ethers = require('ethers');
-const rDAI = require('../src/contracts/rDai.json')
+contract('TESTING', async (accounts) => {
 
-const provider = new ethers.providers.JsonRpcProvider("http://localhost:8545")
+    before(async() => {
+      console.log(accounts[0])
+    });
 
-const contract = new ethers.Contract(
-                    "0xeA718E4602125407fAfcb721b7D760aD9652dfe7",
-                    rDAI,
-                    provider.getSigner(0)
-                  )
-
-async function getHat() {
-  let val = await contract.getHatByAddress("0xb893D8F6779842959C1dfC3095b1c62ceAA16703")
-  console.log(val)
-}
-
-getHat()
-
+    describe("Test Describe", async() => {
+      it("Test it", async () => {
+        assert.equal(true, true, "did not pass")
+      });
+  });
+});
